+++ src/models/relations/BelongsToManyRelation.ts
@@ -1,6 +1,6 @@
 import MultiModelRelation from '@/models/relations/MultiModelRelation';
 import type { ModelConstructor } from '@/models/inference';
-import type { Model } from '@/models/Model';
+import type { Key, Model } from '@/models/Model';
 
 export default class BelongsToManyRelation<
     Parent extends Model = Model,
@@ -8,21 +8,29 @@ export default class BelongsToManyRelation<
     RelatedClass extends ModelConstructor<Related> = ModelConstructor<Related>,
 > extends MultiModelRelation<Parent, Related, RelatedClass> {
 
-    public async resolve(): Promise<Related[]> {
-        const foreignKeys = this.parent.getAttribute<string[]>(this.foreignKeyName);
+    public isEmpty(): boolean {
+        return this.parent.getAttribute<Key[]>(this.foreignKeyName).length === 0;
+    }
 
-        this.related = this.localKeyName === this.relatedClass.primaryKey
-            ? await this.relatedClass.all({
-                $in: foreignKeys,
-            })
-            : await this.relatedClass.all({
-                [this.localKeyName]: {
-                    $or: [
-                        ...foreignKeys.map(key => ({ $eq: key })),
-                        ...foreignKeys.map(key => ({ $contains: key })),
-                    ],
-                },
-            });
+    public async resolve(): Promise<Related[]> {
+        const foreignKeys = this.parent.getAttribute<Key[]>(this.foreignKeyName);
+
+        this.related = this.isEmpty()
+            ? []
+            : (
+                this.localKeyName === this.relatedClass.primaryKey
+                    ? await this.relatedClass.all({
+                        $in: foreignKeys,
+                    })
+                    : await this.relatedClass.all({
+                        [this.localKeyName]: {
+                            $or: [
+                                ...foreignKeys.map(key => ({ $eq: key })),
+                                ...foreignKeys.map(key => ({ $contains: key })),
+                            ],
+                        },
+                    })
+            );
 
         return this.related;
     }